Tests: Move StimulusOntology tests to separate file
